updates on city

+++ src/main/java/com/tms/service/impl/CityServiceImpl.java
@@ -1,11 +1,9 @@
 package com.tms.service.impl;
 
-import org.apache.tomcat.util.net.openssl.ciphers.Cipher;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.tms.common.CommonConstants.City;
-import com.tms.common.CommonConstants.Hotel;
 import com.tms.dao.CityDao;
 import com.tms.model.CityDetails;
 import com.tms.model.TMSResponse;
@@ -20,10 +18,14 @@ public class CityServiceImpl implements CityService {
 	@Override
 	public TMSResponse saveCityDetails(CityDetails cityDetails) {
 		TMSResponse response = new TMSResponse();
+		if (cityDetails.getCityName() == null) {
+			response.setDetails("Invalid request parameters" + cityDetails.toString());
+			return response;
+		}
 		try {
 			CityDetails cityNameObj = cityDao.findByCityName(cityDetails.getCityName());
-			if (cityNameObj != null && cityDetails.getCityName() != null && cityNameObj.getCityName().equalsIgnoreCase(cityDetails.getCityName())) {
-				// given city name should not available in database.
+			if (cityNameObj != null && cityNameObj.getCityName() != null && cityDetails.getCityName() != null
+					&& cityNameObj.getCityName().equalsIgnoreCase(cityDetails.getCityName())) {
 				response.setDetails(City.CITYEXIST);
 			} else {
 				cityDetails.setActive(true);
@@ -38,7 +40,6 @@ public class CityServiceImpl implements CityService {
 		}
 		return response;
 	}
-	// 
 
 	@Override
 	public TMSResponse getCityDetailsById(Integer cityId) {
@@ -60,16 +61,19 @@ public class CityServiceImpl implements CityService {
 	}
 
 	@Override
-	public TMSResponse deleteCityDetailsById(Integer cityId, boolean status) {
+	public TMSResponse deleteCityDetailsById(Integer cityId, Boolean isActive) {
 		TMSResponse response = new TMSResponse();
 		try {
 			CityDetails cityDetailsByIdobj = cityDao.findByCityId(cityId);
 			if (cityDetailsByIdobj != null) {
-				cityDao.deleteByCityId(cityId);
-				if (status) {
-					cityDetailsByIdobj.setActive(status);
+				isActive = isActive == null ? false : isActive;	//if isActive is null then make it as false else take it as isActive 
+				if (isActive) {
+					cityDao.deleteByCityId(cityId);
+				} else {
+					cityDetailsByIdobj.setActive(isActive);
 					response.setData(cityDao.save(cityDetailsByIdobj));
 				}
+
 				response.setDetails(City.DELETED);
 			} else {
 				response.setDetails(City.CITYNOTEXIST);
